pyproject.toml
[project]
name = "nano-banana"
version = "1.0.0"
description = "Image generation CLI for Claude Code using Google Gemini API"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]
